requirements.txt
requests~=2.25.1
beautifulsoup4~=4.9.3
python-dotenv~=0.17.1
discord.py==1.7.2